#patients ownership by dentists configured

+++ src/modules/patients/patients.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { UserEntity } from '@users/user.entity';
 import {
   CreatePatientDto,
   GetPatientsFilterDto,
@@ -15,8 +16,13 @@ export class PatientsService {
     private readonly patientRepository: Repository<PatientEntity>,
   ) {}
 
-  async getPatients(filterDto: GetPatientsFilterDto): Promise<PatientEntity[]> {
+  async getPatients(
+    filterDto: GetPatientsFilterDto,
+    user: UserEntity,
+  ): Promise<PatientEntity[]> {
     const query = this.patientRepository.createQueryBuilder('patient');
+
+    query.andWhere(`patient.userId = :userId`, { userId: user.id });
 
     Object.entries(filterDto).forEach(([key, value]) => {
       if (!value) return;
@@ -29,8 +35,11 @@ export class PatientsService {
     return await query.getMany();
   }
 
-  async getPatientById(id: number): Promise<PatientEntity> {
-    const patient = await this.patientRepository.findOneBy({ id });
+  async getPatientById(id: number, user: UserEntity): Promise<PatientEntity> {
+    const patient = await this.patientRepository.findOneBy({
+      id,
+      userId: user.id,
+    });
 
     if (!patient) {
       throw new NotFoundException(`Patient with id ${id} not found`);
@@ -41,6 +50,7 @@ export class PatientsService {
 
   async createPatient(
     createPatientDto: CreatePatientDto,
+    user: UserEntity,
   ): Promise<PatientEntity> {
     const { firstname, lastname, phone, email, description } = createPatientDto;
 
@@ -51,8 +61,11 @@ export class PatientsService {
     patient.phone = phone;
     patient.email = email;
     patient.description = description;
+    patient.user = user;
 
     await patient.save();
+
+    delete patient.user;
 
     return patient;
   }
@@ -60,21 +73,19 @@ export class PatientsService {
   async updatePatientById(
     id: number,
     updatePatientDto: UpdatePatientDto,
+    user: UserEntity,
   ): Promise<PatientEntity> {
-    const patient = await this.patientRepository.preload({
-      id,
-      ...updatePatientDto,
-    });
+    const patient = await this.getPatientById(id, user);
 
-    if (!patient) {
-      throw new NotFoundException(`Patient with id ${id} not found`);
-    }
+    Object.keys(updatePatientDto).map((key) => {
+      patient[key] = updatePatientDto[key];
+    });
 
     return await this.patientRepository.save(patient);
   }
 
-  async deletePatient(id: number): Promise<void> {
-    const result = await this.patientRepository.delete(id);
+  async deletePatient(id: number, user: UserEntity): Promise<void> {
+    const result = await this.patientRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Patient with id ${id} not found`);
